libreria/requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==3.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.7
dnspython==2.7.0
drf-spectacular==0.28.0
Faker==33.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
PyJWT==2.10.1
pymongo==3.11.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
